+++ Model/GPSHis.cs
@@ -13,9 +13,9 @@ namespace Model
 
         public Guid GPSItemId { get; set; }
 
-        public double Lat { get; set; }
+        public decimal Lat { get; set; }
 
-        public double Lng { get; set; }
+        public decimal Lng { get; set; }
 
         public string LoactionInfo { get; set; }
