+++ Model/GPSItem.cs
@@ -21,9 +21,9 @@ namespace Model
 
         public string CardNum { get; set; }
 
-        public double Lat { get; set; }
+        public decimal Lat { get; set; }
 
-        public double Lng { get; set; }
+        public decimal Lng { get; set; }
 
         public string Speed { get; set; }
